gpu: nvgpu: boardobj: use size_t for construct API
Rather than having everyone cast their sizeof() values to u16, just use
size_t for the size for the boardobj_construct_super() API.

JIRA NVGPU-1008

Reviewed-on: https://git-master.nvidia.com/r/2011433
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/boardobj/boardobj.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016-2018, NVIDIA CORPORATION.  All rights reserved.
+ * Copyright (c) 2016-2019, NVIDIA CORPORATION.  All rights reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -26,7 +26,7 @@
 #include <nvgpu/pmuif/ctrlboardobj.h>
 
 int boardobj_construct_super(struct gk20a *g, struct boardobj **ppboardobj,
-				u16 size, void *args)
+				size_t size, void *args)
 {
 	struct boardobj  *pboardobj = NULL;
 	struct boardobj  *devtmp = (struct boardobj *)args;
